Reposição permite saida, filtro de tempo em historico de remedio e diario, notificações separadas adequadamente por perfil

+++ lib/models/medication_model.dart
@@ -18,7 +18,7 @@ enum MedicationType {
 class Medication {
   final int? id;
   final String nome;
-  final int estoque;
+  final double estoque;
   final MedicationType tipo;
   final bool deletado;
   final int idPerfil;
